backend/requirements.txt
# Core Dependencies 
Flask==3.0.0 
Flask-SQLAlchemy==3.1.1 
Flask-Migrate==4.0.5 
Flask-CORS==4.0.0 
Flask-SocketIO==5.3.5 
python-socketio[client]==5.10.0 
Flask-Bcrypt==1.0.1 
 
# Database 
psycopg2-binary==2.9.9 
SQLAlchemy==2.0.23 
alembic==1.13.1 
 
# Authentication
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3
cryptography==41.0.7 
 
# Redis 
redis==5.0.1 
hiredis==2.2.3 
 
# Twilio 
twilio==8.10.3 
 
# Server 
gunicorn==21.2.0 
eventlet==0.33.3 
 
# Utilities
python-dotenv==1.0.0
Faker==20.1.0

# HTTP Requests 
requests==2.31.0 
marshmallow==3.20.1 
python-dateutil==2.8.2 
 
# Testing and Linting 
pytest==7.4.3 
pytest-flask==1.3.0 
black==23.12.1 
flake8==7.0.0 
